refactor llm declaration
