+++ src/llm/agents/reactAgent.ts
@@ -1,10 +1,7 @@
-import { ChatOpenAI } from "@langchain/openai";
 import { createReactAgent } from "@langchain/langgraph/prebuilt";
 import { createBrowserTools } from "../tools/browser/browserTools";
-import { storageService } from "@/storage/services/storageService";
-import { Settings } from "@/storage/types/settings";
-import { LLMProviders } from "../types/providers";
 import { browserStateService } from "@/browser/services/browserStateService";
+import { createLLM } from "../services/llmService";
 
 export const BrowserResponseSchema = {
     type: "object",
@@ -18,28 +15,7 @@ export const BrowserResponseSchema = {
 };
 
 export async function createAgent() {
-    const response = await storageService.getData("settings");
-    if (!response.success || !response.data)
-    {
-        throw new Error('Failed to load settings');
-    }
-
-    const settings = response.data as Settings;
-    const openAIKey = settings.apiKeys.find(
-        key => key.provider.toLowerCase() === LLMProviders.OPENAI.toLowerCase()
-    )?.key;
-
-    if (!openAIKey)
-    {
-        throw new Error('OpenAI API key not configured');
-    }
-
-    const model = new ChatOpenAI({
-        openAIApiKey: openAIKey,
-        modelName: "gpt-4o-mini",
-        temperature: 0,
-        streaming: true
-    });
+    const model = await createLLM();
 
     const browserTools = createBrowserTools();
     const llmWithTools = model.bindTools(browserTools, { parallel_tool_calls: false });
