+++ src/llm/agents/verifierAgent.ts
@@ -1,33 +1,9 @@
-import { ChatOpenAI } from "@langchain/openai";
 import { createReactAgent } from "@langchain/langgraph/prebuilt";
-import { storageService } from "@/storage/services/storageService";
 import { browserStateService } from "@/browser/services/browserStateService";
-import { Settings } from "@/storage/types/settings";
-import { LLMProviders } from "../types/providers";
+import { createLLM } from "../services/llmService";
 
 export async function createVerifierAgent() {
-    const response = await storageService.getData("settings");
-    if (!response.success || !response.data)
-    {
-        throw new Error('Failed to load settings');
-    }
-
-    const settings = response.data as Settings;
-    const openAIKey = settings.apiKeys.find(
-        key => key.provider.toLowerCase() === LLMProviders.OPENAI.toLowerCase()
-    )?.key;
-
-    if (!openAIKey)
-    {
-        throw new Error('OpenAI API key not configured');
-    }
-
-    const model = new ChatOpenAI({
-        openAIApiKey: openAIKey,
-        modelName: "gpt-4o-mini",
-        temperature: 0,
-        streaming: true
-    });
+    const model = await createLLM();
 
     const browserState = await browserStateService.getBrowserState();
     const stateMessage = browserState.success ? browserState.message : 'Failed to get browser state';
