--- src/llm/services/openai/openaiService.ts
@@ -1,59 +1,0 @@
-import { BaseLLMService } from "../baseLLMService";
-import { LLMRequest, LLMResponse } from "../../types";
-
-/**
- * OpenAI service implementation
- * Note: This is a placeholder that will be implemented when OpenAI integration is added
- */
-export class OpenAIService extends BaseLLMService {
-    /**
-     * Process text using OpenAI
-     * @param {LLMRequest} request - The request to process
-     * @returns {Promise<LLMResponse>} The OpenAI response
-     */
-    async process(_request: LLMRequest): Promise<LLMResponse> {
-        if (!this.validateConfig())
-        {
-            return this.createErrorResponse("Invalid OpenAI configuration");
-        }
-
-        try
-        {
-            // Placeholder for OpenAI API integration
-            // Will be implemented when adding OpenAI support
-            return this.createSuccessResponse(
-                "OpenAI integration not yet implemented",
-                0,
-                { provider: "openai" }
-            );
-        } catch (error)
-        {
-            return this.createErrorResponse(
-                error instanceof Error ? error : String(error)
-            );
-        }
-    }
-
-    /**
-     * Initialize OpenAI service
-     * @returns {Promise<boolean>} Success status
-     */
-    async initialize(): Promise<boolean> {
-        if (!this.validateConfig())
-        {
-            return false;
-        }
-
-        // Placeholder for OpenAI initialization
-        // Will be implemented when adding OpenAI support
-        return true;
-    }
-
-    /**
-     * Validate OpenAI specific configuration
-     * @returns {boolean} Is configuration valid
-     */
-    protected validateConfig(): boolean {
-        return super.validateConfig() && this.config.provider === "openai";
-    }
-}
